requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
deprecation==2.1.0
Django==5.0.1
ecdsa==0.18.0
idna==3.6
packaging==23.2
pyasn1==0.5.1
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==3.7.0
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
